+++ decoder/htj2k/htj2k_roundtrip_test.go
@@ -24,7 +24,7 @@ func TestRoundTripRGBA(t *testing.T) {
 		}
 	}
 
-	cs, err := encodeTestLossless(orig, w, h)
+	cs, err := encodeTestLossless(orig, w, h, 1)
 	if err != nil {
 		t.Fatalf("encodeTestLossless: %v", err)
 	}
@@ -86,7 +86,7 @@ func TestRoundTripRGBAWithDst(t *testing.T) {
 		}
 	}
 
-	cs, err := encodeTestLossless(orig, w, h)
+	cs, err := encodeTestLossless(orig, w, h, 1)
 	if err != nil {
 		t.Fatalf("encodeTestLossless: %v", err)
 	}